+++ Installomator.sh
@@ -71,6 +71,11 @@ INSTALL=""
 #                  if it is newer/different in version
 #  - force         Install even if it’s the same version
 
+# timeout for dialog box in seconds
+# prevents the script from blocking other processes while the user ignores the
+# dialogue box for a prolonged period of time.
+# TIMEOUT=300 # wait 5 minutes for user to confirm before giving up
+TIMEOUT=
 
 # NOTE: How labels work
 
@@ -189,16 +194,79 @@ runAsUser() {
     fi
 }
 
+# return a human readable text for the TIMEOUT value
+timeout_text() {
+    if [[ -z $TIMEOUT ]]; then
+        echo ""
+        return
+    elif [[ $TIMEOUT -lt 60 ]]; then
+        echo "$TIMEOUT seconds"
+        return
+    elif [[ $TIMEOUT -lt 3600 ]]; then
+        minutes=$(( $TIMEOUT / 60 ))
+        minutes_text="$minutes minute"
+
+        if [[ $minutes != 1 ]]; then
+            minutes_text="${minutes_text}s"
+        fi
+
+        seconds=$(( $TIMEOUT % 60 ))
+        seconds_text=""
+
+        if [[ $seconds -gt 0 ]]; then
+            seconds_text=" and $seconds second"
+
+            if [[ $seconds != 1 ]]; then
+                seconds_text="${seconds_text}s"
+            fi
+        fi
+
+        echo "${minutes_text}${seconds_text}"
+        return
+    fi
+
+    hours=$(( $TIMEOUT / 60 / 60 ))
+    hours_text="$hours hour"
+
+    if [[ $hours != 1 ]]; then
+        hours_text="${hours_text}s"
+    fi
+
+    minutes=$(( $TIMEOUT / 60 % 60 ))
+    minutes_text=""
+
+    if [[ $minutes -gt 0 ]]; then
+        minutes_text=" and $minutes minute"
+
+        if [[ $minutes != 1 ]]; then
+            minutes_text="${minutes_text}s"
+        fi
+    fi
+
+    echo "${hours_text}${minutes_text}"
+}
+
 displaydialog() { # $1: message $2: title
     message=${1:-"Message"}
     title=${2:-"Installomator"}
-    runAsUser osascript -e "button returned of (display dialog \"$message\" with  title \"$title\" buttons {\"Not Now\", \"Quit and Update\"} default button \"Quit and Update\" with icon POSIX file \"$LOGO\")"
+    giving_up=
+
+    if [ -n "$TIMEOUT" ]; then
+        giving_up=" giving up after $TIMEOUT"
+    fi
+
+    runAsUser osascript -e "button returned of (display dialog \"$message\" with  title \"$title\" buttons {\"Not Now\", \"Quit and Update\"} default button \"Quit and Update\" $giving_up with icon POSIX file \"$LOGO\")"
 }
 
 displaydialogContinue() { # $1: message $2: title
     message=${1:-"Message"}
     title=${2:-"Installomator"}
-    runAsUser osascript -e "button returned of (display dialog \"$message\" with  title \"$title\" buttons {\"Quit and Update\"} default button \"Quit and Update\" with icon POSIX file \"$LOGO\")"
+
+    if [ -n "$TIMEOUT" ]; then
+        giving_up=" giving up after $TIMEOUT"
+    fi
+
+    runAsUser osascript -e "button returned of (display dialog \"$message\" with  title \"$title\" buttons {\"Quit and Update\"} default button \"Quit and Update\" $giving_up with icon POSIX file \"$LOGO\")"
 }
 
 displaynotification() { # $1: message $2: title
@@ -339,18 +407,25 @@ checkRunningProcesses() {
         for x in ${blockingProcesses}; do
             if pgrep -xq "$x"; then
                 printlog "found blocking process $x"
-                appClosed=1
                 
                 case $BLOCKING_PROCESS_ACTION in
                     kill)
                       printlog "killing process $x"
                       pkill $x
+                      appClosed=1
                       sleep 5
                       ;;
                     prompt_user|prompt_user_then_kill)
-                      button=$(displaydialog "Quit “$x” to continue updating? (Leave this dialogue if you want to activate this update later)." "The application “$x” needs to be updated.")
+                      if [ -n "$TIMEOUT" ]; then
+                        timer_message="(This dialogue will close in $(timeout_text))"
+                      else
+                        timer_message="(Leave this dialogue if you want to activate this update later)"
+                      fi
+                      button=$(displaydialog "Quit “$x” to continue updating? $timer_message." "The application “$x” needs to be updated.")
                       if [[ $button = "Not Now" ]]; then
                         cleanupAndExit 10 "user aborted update"
+                      elif [[ $button = "" ]]; then
+                        cleanupAndExit 13 "dialog timed out after $(timeout_text)"
                       else
                         if [[ $i > 2 && $BLOCKING_PROCESS_ACTION = "prompt_user_then_kill" ]]; then
                           printlog "Changing BLOCKING_PROCESS_ACTION to kill"
@@ -358,6 +433,7 @@ checkRunningProcesses() {
                         else
                           printlog "telling app $x to quit"
                           runAsUser osascript -e "tell app \"$x\" to quit"
+                          appClosed=1
                           # give the user a bit of time to quit apps
                           printlog "waiting 30 seconds for processes to quit"
                           sleep 30
@@ -365,10 +441,15 @@ checkRunningProcesses() {
                       fi
                       ;;
                     prompt_user_loop)
-                      button=$(displaydialog "Quit “$x” to continue updating? (Click “Not Now” to be asked in 1 hour, or leave this open until you are ready)." "The application “$x” needs to be updated.")
-                      if [[ $button = "Not Now" ]]; then
+                      if [ -n "$TIMEOUT" ]; then
+                        timer_message="(Click “Not Now” to be asked in 1 hour)"
+                      else
+                        timer_message="(Click “Not Now” to be asked in 1 hour, or leave this open until you are ready)"
+                      fi
+                      button=$(displaydialog "Quit “$x” to continue updating? $timer_message." "The application “$x” needs to be updated.")
+                      if [[ $button = "Not Now" || $button = "" ]]; then
                         if [[ $i < 2 ]]; then
-                          printlog "user wants to wait an hour"
+                          printlog "user clicked not now or timed out after $(timeout_text), waiting for an hour"
                           sleep 3600 # 3600 seconds is an hour
                         else
                           printlog "change of BLOCKING_PROCESS_ACTION to tell_user"
@@ -377,15 +458,22 @@ checkRunningProcesses() {
                       else
                         printlog "telling app $x to quit"
                         runAsUser osascript -e "tell app \"$x\" to quit"
+                        appClosed=1
                         # give the user a bit of time to quit apps
                         printlog "waiting 30 seconds for processes to quit"
                         sleep 30
                       fi
                       ;;
                     tell_user|tell_user_then_kill)
-                      button=$(displaydialogContinue "Quit “$x” to continue updating? (This is an important update). Wait for notification of update before launching app again." "The application “$x” needs to be updated.")
+                      if [ -n "$TIMEOUT" ]; then
+                        timer_message="(This important update will automatically proceed in $(timeout_text))"
+                      else
+                        timer_message="(This is an important update)"
+                      fi
+                      button=$(displaydialogContinue "Quit “$x” to continue updating? $timer_message. Wait for notification of update before launching app again." "The application “$x” needs to be updated.")
                       printlog "telling app $x to quit"
                       runAsUser osascript -e "tell app \"$x\" to quit"
+                      appClosed=1
                       # give the user a bit of time to quit apps
                       printlog "waiting 30 seconds for processes to quit"
                       sleep 30